+++ clowder/utility/file_system.py
