+++ clowder/utility/git/git_repo.py
@@ -9,7 +9,7 @@ import sys
 from git import Repo, GitError
 from termcolor import colored, cprint
 
-from clowder.utility.clowder_utilities import remove_directory
+from clowder.utility.file_system import remove_directory
 from clowder.utility.connectivity import is_offline
 from clowder.utility.error.clowder_git_error import ClowderGitError
 from clowder.utility.execute import execute_command
